Backend/develop

See merge request s09-final/S09P31E203!110

+++ backend/timing/src/main/java/com/kkukku/timing/apis/challenge/responses/ChallengeResponse.java
@@ -20,8 +20,8 @@ public class ChallengeResponse {
     public static class Challenge {
 
         private String thumbnailUrl;
-        private int countDays;
-        private int maxDays;
+        private long countDays;
+        private long maxDays;
 
     }
 }
